27 check entity existence before deleting (#48)
* Added model path field

* Fixed deleting issues

---------

+++ backend/routers/LossFunctionsRouter.py
@@ -39,13 +39,21 @@ def add_loss_function(loss_function_body: LossFunctionSummary):
 
 @router.delete("/{loss_function_id}", status_code=status.HTTP_200_OK)
 def delete_loss_function(loss_function_id: int):
-    try:
-        cursor.execute("delete from loss_function where id = ?", (loss_function_id,))
-        logging.info(f"Loss function with id = {str(loss_function_id)} has been deleted successfully")
-    except mariadb.Error as e:
-        logging.error(f"Could not delete loss function with id = {str(loss_function_id)}. Error: {e}")
-        return JSONResponse(status_code=status.HTTP_400_BAD_REQUEST,
-                            content=f"Could not delete loss function with id = {str(loss_function_id)}")
+    get_response = get_loss_function(loss_function_id)
+    if get_response.status_code == status.HTTP_200_OK:
+        try:
+            cursor.execute("delete from loss_function where id = ?", (loss_function_id,))
+            connection.commit()
+            logging.info(f"Loss function with id = {str(loss_function_id)} has been deleted successfully")
+        except mariadb.Error as e:
+            logging.error(f"Could not delete loss function with id = {str(loss_function_id)}. Error: {e}")
+            return JSONResponse(status_code=status.HTTP_400_BAD_REQUEST,
+                                content=f"Could not delete loss function with id = {str(loss_function_id)}")
+    else:
+        logging.error(f"Could not delete loss function with id = {str(loss_function_id)}. Error: entity does not exist.")
+        JSONResponse(status_code=status.HTTP_400_BAD_REQUEST,
+                     content=f"Could not delete loss function with id = {str(loss_function_id)} because entity does not "
+                             f"exist.")
 
 
 @router.get("/", status_code=status.HTTP_200_OK)
